requirements-dev.txt
flake8
coverage
Sphinx
pytest>=2.9
pytest-xdist
pytest-cov
pytest-django
mock==2.0.0
responses==0.5.1
